backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pydantic==2.7.1
pydantic-settings==2.7.1
openai==1.109.1
pydub==0.25.1
python-dotenv==1.0.0
httpx==0.27.0
slowapi==0.1.8
aiofiles==23.2.1
email-validator==2.1.0.post1
google-auth==2.28.1
requests==2.31.0
PyJWT==2.8.0
# Object storage (R2 / MinIO / S3)
aioboto3==12.3.0
# Supabase JWT verification handled by PyJWT (no SDK needed)
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
aiosqlite==0.19.0
